Changed nanoTime() double to long

+++ src/by/bsu/parallel/main/Main.java
@@ -22,27 +22,27 @@ public class Main {
         Random rand = new Random();
 
         System.out.println("Initializing matrix...");
-        double startInit = System.nanoTime();
+        long startInit = System.nanoTime();
         for (int i = 0; i < SIZE; i++) {
             for (int j = 0; j < SIZE; j++) {
                 matrix[i][j] = rand.nextInt();
             }
         }
-        double finishInit = System.nanoTime();
+        long finishInit = System.nanoTime();
         System.out.println("POSLEDOV " + (finishInit - startInit));
 
 
-        double start4Init = System.nanoTime();
+        long start4Init = System.nanoTime();
         WorkController workControllerInit = new WorkController(4);
         workControllerInit.doWorkMatrix(new InitializeMatrix(), matrix2);
-        double finish4Init = System.nanoTime();
+        long finish4Init = System.nanoTime();
         System.out.println("2 POTOKA " + (finish4Init - start4Init));
 
         System.out.println("Starting benchmark...");
         System.out.println();
 
         System.out.println("-=-=-Multiplication-=-=-");
-        double start4 = System.nanoTime();
+        long start4 = System.nanoTime();
         for (int i = 0; i < mulResult.length; i++) { // aRow
             for (int j = 0; j < mulResult.length; j++) { // bColumn
                 for (int k = 0; k < mulResult.length; k++) { // aColumn
@@ -50,17 +50,17 @@ public class Main {
                 }
             }
         }
-        double finish4 = System.nanoTime();
+        long finish4 = System.nanoTime();
         System.out.println("POSLEDOV " + (finish4 - start4));
         for (int i = 0; i < SIZE; i++) {
             for (int j = 0; j < SIZE; j++) {
                 mulResult[i][j] = 0;
             }
         }
-        double start4Mult = System.nanoTime();
+        long start4Mult = System.nanoTime();
         WorkController workControllerMult = new WorkController(2);
         workControllerMult.doWorkMatrix(new MatrixMultiply(), matrix, matrix2);
-        double finish4Mult = System.nanoTime();
+        long finish4Mult = System.nanoTime();
         System.out.println("2 POTOKA " + (finish4Mult - start4Mult));
         for (int i = 0; i < SIZE; i++) {
             for (int j = 0; j < SIZE; j++) {
@@ -68,10 +68,10 @@ public class Main {
             }
         }
 
-        double start4Mult3 = System.nanoTime();
+        long start4Mult3 = System.nanoTime();
         WorkController workControllerMult3 = new WorkController(3);
         workControllerMult3.doWorkMatrix(new MatrixMultiply(), matrix, matrix2);
-        double finish4Mult3 = System.nanoTime();
+        long finish4Mult3 = System.nanoTime();
         System.out.println("3 POTOKA " + (finish4Mult3 - start4Mult3));
         for (int i = 0; i < SIZE; i++) {
             for (int j = 0; j < SIZE; j++) {
@@ -79,16 +79,16 @@ public class Main {
             }
         }
 
-        double start4Mult2 = System.nanoTime();
+        long start4Mult2 = System.nanoTime();
         WorkController workControllerMult2 = new WorkController(4);
         workControllerMult2.doWorkMatrix(new MatrixMultiply(), matrix, matrix2);
-        double finish4Mult2 = System.nanoTime();
+        long finish4Mult2 = System.nanoTime();
         System.out.println("4 POTOKA " + (finish4Mult2 - start4Mult2));
 
         ///////////////////////////////////////////////////////////////////
         System.out.println();
         System.out.println("-=-=-Max Line Element-=-=-");
-        double start2 = System.nanoTime();
+        long start2 = System.nanoTime();
         double max;
         for (int i = 0; i < SIZE; i++) {
             max = Double.MIN_VALUE;
@@ -99,31 +99,31 @@ public class Main {
             }
             result[i] = max;
         }
-        double finish2 = System.nanoTime();
+        long finish2 = System.nanoTime();
         System.out.println("POSLEDOV " + (finish2 - start2));
 
-        double start = System.nanoTime();
+        long start = System.nanoTime();
         WorkController workController = new WorkController(2);
         workController.doWorkVector(new MaxLineElement(), matrix);
-        double finish = System.nanoTime();
+        long finish = System.nanoTime();
         System.out.println("2 POTOKA " + (finish - start));
         for (int i = 0; i < result.length; i++) {
             result[i] = 0;
         }
 
-        double start3 = System.nanoTime();
+        long start3 = System.nanoTime();
         WorkController workController3 = new WorkController(3);
         workController3.doWorkVector(new MaxLineElement(), matrix);
-        double finish3 = System.nanoTime();
+        long finish3 = System.nanoTime();
         System.out.println("3 POTOKA " + (finish3 - start3));
         for (int i = 0; i < result.length; i++) {
             result[i] = 0;
         }
 
-        double start45 = System.nanoTime();
+        long start45 = System.nanoTime();
         WorkController workController4 = new WorkController(3);
         workController4.doWorkVector(new MaxLineElement(), matrix);
-        double finish45 = System.nanoTime();
+        long finish45 = System.nanoTime();
         System.out.println("4 POTOKA " + (finish45 - start45));
         for (int i = 0; i < result.length; i++) {
             result[i] = 0;
@@ -132,38 +132,38 @@ public class Main {
         ///////////////////////////////////////////////////////////
         System.out.println();
         System.out.println("-=-=-Scalar Vector Multiply-=-=-");
-        double startOdno = System.nanoTime();
+        long startOdno = System.nanoTime();
         for (int i = 0; i < SIZE; i++) {
             for (int j = 0; j < SIZE; j++) {
                 result[i] += matrix[i][j] * matrix2[i][j];
             }
         }
-        double finishOdno = System.nanoTime();
+        long finishOdno = System.nanoTime();
         System.out.println("POSLEDOV " + (finishOdno - startOdno));
 
-        double startZero = System.nanoTime();
+        long startZero = System.nanoTime();
         WorkController workControllerz = new WorkController(2);
         ScalarVectorMultiply scalarVectorMultiply = new ScalarVectorMultiply();
         workControllerz.doWorkMatrix(scalarVectorMultiply, matrix, matrix2);
-        double finishZero = System.nanoTime();
+        long finishZero = System.nanoTime();
         System.out.println("2 POTOKA " + (finishZero - startZero));
         for (int i = 0; i < result.length; i++) {
             result[i] = 0;
         }
 
-        double startOne = System.nanoTime();
+        long startOne = System.nanoTime();
         WorkController workControllerOne = new WorkController(3);
         workControllerOne.doWorkMatrix(scalarVectorMultiply, matrix, matrix2);
-        double finishOne = System.nanoTime();
+        long finishOne = System.nanoTime();
         System.out.println("3 POTOKA " + (finishOne - startOne));
         for (int i = 0; i < result.length; i++) {
             result[i] = 0;
         }
 
-        double startTwo = System.nanoTime();
+        long startTwo = System.nanoTime();
         WorkController workControllerTwo = new WorkController(4);
         workControllerTwo.doWorkMatrix(scalarVectorMultiply, matrix, matrix2);
-        double finishTwo = System.nanoTime();
+        long finishTwo = System.nanoTime();
         System.out.println("4 POTOKA " + (finishTwo - startTwo));
 
         System.out.println();
